Move dispersion-estimation API out of the initial bundle
CI's initial-gzip budget tripped at 244.1 KiB vs 244: the dispersion
client functions and their runtime parsers landed in api/client.ts and
types/guards.ts, which ship in the initial chunk, though their only
consumer is the lazy-loaded modelling config panel. They now live in
api/dispersion.ts (with lean local parsers), pulled in via the panel's
lazy chunk instead — initial total back to 243.2 KiB. client.ts exports
its request/post fetch machinery for split-chunk API modules.
